+++ packages/components/src/components/Icons/NewsIcon.tsx
@@ -1,7 +1,8 @@
-import Icon from '../Icon/Icon';
+import React from 'react';
+import { Icon } from '../Icon/Icon';
 import type { IconPropsWithoutChildren } from '../Icon/Icon';
 
-const News = (props: IconPropsWithoutChildren) => {
+export const NewsIcon = (props: IconPropsWithoutChildren) => {
   return (
     <Icon {...props}>
       <svg
@@ -25,5 +26,3 @@ const News = (props: IconPropsWithoutChildren) => {
     </Icon>
   );
 };
-
-export default News;