+++ packages/components/src/components/Icons/SubtitleIcon.tsx
@@ -1,7 +1,8 @@
-import Icon from '../Icon/Icon';
+import React from 'react';
+import { Icon } from '../Icon/Icon';
 import type { IconPropsWithoutChildren } from '../Icon/Icon';
 
-const Subtitle = (props: IconPropsWithoutChildren) => {
+export const SubtitleIcon = (props: IconPropsWithoutChildren) => {
   return (
     <Icon {...props}>
       <svg
@@ -20,5 +21,3 @@ const Subtitle = (props: IconPropsWithoutChildren) => {
     </Icon>
   );
 };
-
-export default Subtitle;