+++ packages/volto/src/components/manage/Preferences/PersonalPreferences.test.jsx
@@ -8,9 +8,7 @@ import PersonalPreferences from './PersonalPreferences';
 
 const mockStore = configureStore();
 
-jest.mock('react-portal', () => ({
-  Portal: jest.fn(() => <div id="Portal" />),
-}));
+jest.mock('../Toolbar/Toolbar', () => jest.fn(() => <div id="Portal" />));
 
 jest.mock('@plone/volto/helpers/Loadable/Loadable');
 beforeAll(
@@ -29,6 +27,13 @@ describe('PersonalPreferences', () => {
         'plone.app.vocabularies.Keywords': {
           items: [{ title: 'My item', value: 'myitem' }],
           itemsTotal: 1,
+        },
+      },
+      content: {
+        data: {},
+        create: {
+          loading: false,
+          loaded: true,
         },
       },
     });